requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.1
solders==0.18.1
base58==2.1.1
selenium==4.15.2
pytesseract==0.3.10
Pillow==10.1.0
base58
flask
Pillow
pytesseract
python-dotenv
requests
selenium
solders
bip-utils